pyproject.toml
[project]
name = "random-crate"
version = "0.1.0"
description = "Add your description here"
license = "MIT License"
readme = "README.md"
authors = [
  {name = "Janusz \"Ivellios\" Kamieński"},
]
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipython>=9.0.2",
]
